+++ Server/src/GameWorld/Weapon/Scout.h
@@ -9,7 +9,7 @@ class Scout : public Weapon {
     public:
         Scout();
         void shoot(std::shared_ptr<Collidable>& player, bool right,
-                   std::map<int, std::shared_ptr<Infected>>& infecteds) override;
+                   std::map<uint8_t, std::shared_ptr<Infected>>& infecteds) override;
 };
